refactor(client): extract the record drawer into its own OnPush component (A17.9b-5)
The detail drawer edits one memory/entity/edge/chrono record and is opened from
every record tab. It was ~290 lines of inline template plus its open/save/close
cycle living in the shell. With the picker and shared helpers already split out
(9b-3/9b-4), the drawer's every dependency is a service or util, so it can finally
own its state and view:

- record-drawer-state.service.ts (RecordDrawerState): the drawer signals, the four
  plain edit models, and open/save/close (+ the drawer branch of onEntityTypeChange),
  consuming BrainStore + EntityRefPicker + BrainApi and the brain-format utils.
  spaceId is fed from the shell (nav state stays on the shell). Bodies moved verbatim
  (openDrawer->open, closeDrawer->close, saveDrawer->save; activeSpaceId()->spaceId()).
- record-drawer.component.ts (RecordDrawerComponent): OnPush, standalone; the ~290-line
  drawer template, injecting the state + store + picker. The shell renders
  <app-record-drawer /> and the tab view-detail buttons call drawerState.open(...).
- brain-form.styles.ts: the chip/flyout + drawer CSS, shared by the shell's forms and
  the drawer (Angular scopes styles per component, so both source them from one const).
- chronoStatusOptions joined chronoKinds on BrainStore; dead drawerEditFileMeta removed.

The drawer's OnPush contract is load-bearing (plain ngModel models render only because
open() writes the drawerRecord signal in the same turn); the two rendering tests that
pin it, plus an OnPush assertion, moved to record-drawer.component.spec.ts and drive
the component directly.

brain.component.ts 3349 -> 2845 (the shell has shed ~860 lines across A17.9b). Client
suite 168 -> 169. tsc --noUnusedLocals clean on the brain files; prod AOT build
warning-free bar the pre-existing qrcode CJS bailout.
